+++ tests/schema/test_generate_schema.py
@@ -85,7 +85,7 @@ class TestGenerateSchema(test.SimpleTestCase):
         with self.assertRaisesRegex(
             ConfigurationError, "Can't create schema due to cyclic fk references"
         ):
-            await self.init_for("tests.models_cyclic")
+            await self.init_for("tests.schema.models_cyclic")
 
     async def test_create_index(self):
         await self.init_for("tests.testmodels")
@@ -96,25 +96,37 @@ class TestGenerateSchema(test.SimpleTestCase):
         with self.assertRaisesRegex(
             ConfigurationError, 'Foreign key accepts model name in format "app.Model"'
         ):
-            await self.init_for("tests.models_fk_1")
+            await self.init_for("tests.schema.models_fk_1")
 
     async def test_fk_bad_on_delete(self):
         with self.assertRaisesRegex(
             ConfigurationError, "on_delete can only be CASCADE, RESTRICT or SET_NULL"
         ):
-            await self.init_for("tests.models_fk_2")
+            await self.init_for("tests.schema.models_fk_2")
 
     async def test_fk_bad_null(self):
         with self.assertRaisesRegex(
             ConfigurationError, "If on_delete is SET_NULL, then field must have null=True set"
         ):
-            await self.init_for("tests.models_fk_3")
+            await self.init_for("tests.schema.models_fk_3")
+
+    async def test_o2o_bad_on_delete(self):
+        with self.assertRaisesRegex(
+            ConfigurationError, "on_delete can only be CASCADE, RESTRICT or SET_NULL"
+        ):
+            await self.init_for("tests.schema.models_o2o_2")
+
+    async def test_o2o_bad_null(self):
+        with self.assertRaisesRegex(
+            ConfigurationError, "If on_delete is SET_NULL, then field must have null=True set"
+        ):
+            await self.init_for("tests.schema.models_o2o_3")
 
     async def test_m2m_bad_model_name(self):
         with self.assertRaisesRegex(
             ConfigurationError, 'Foreign key accepts model name in format "app.Model"'
         ):
-            await self.init_for("tests.models_m2m_1")
+            await self.init_for("tests.schema.models_m2m_1")
 
     async def test_table_and_row_comment_generation(self):
         await self.init_for("tests.testmodels")
@@ -125,7 +137,7 @@ class TestGenerateSchema(test.SimpleTestCase):
 
     async def test_schema(self):
         self.maxDiff = None
-        await self.init_for("tests.models_schema_create")
+        await self.init_for("tests.schema.models_schema_create")
         sql = get_schema_sql(Tortoise.get_connection("default"), safe=False)
         self.assertEqual(
             sql.strip(),
@@ -155,6 +167,12 @@ CREATE TABLE "team" (
 ) /* The TEAMS! */;
 CREATE INDEX "idx_team_manager_676134" ON "team" ("manager_id", "key");
 CREATE INDEX "idx_team_manager_ef8f69" ON "team" ("manager_id", "name");
+CREATE TABLE "teamaddress" (
+    "city" VARCHAR(50) NOT NULL  /* City */,
+    "country" VARCHAR(50) NOT NULL  /* Country */,
+    "street" VARCHAR(128) NOT NULL  /* Street Address */,
+    "team_id" VARCHAR(50) NOT NULL UNIQUE PRIMARY KEY REFERENCES "team" ("name") ON DELETE CASCADE
+);
 CREATE TABLE "tournament" (
     "tid" INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
     "name" VARCHAR(100) NOT NULL  /* Tournament name */,
@@ -169,9 +187,16 @@ CREATE TABLE "event" (
     "token" VARCHAR(100) NOT NULL UNIQUE /* Unique token */,
     "key" VARCHAR(100) NOT NULL,
     "tournament_id" SMALLINT NOT NULL REFERENCES "tournament" ("tid") ON DELETE CASCADE /* FK to tournament */,
-    UNIQUE ("name", "prize"),
-    UNIQUE ("tournament_id", "key")
+    CONSTRAINT "uid_event_name_c6f89f" UNIQUE ("name", "prize"),
+    CONSTRAINT "uid_event_tournam_a5b730" UNIQUE ("tournament_id", "key")
 ) /* This table contains a list of all the events */;
+CREATE TABLE "venueinformation" (
+    "id" INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
+    "name" VARCHAR(128) NOT NULL,
+    "capacity" INT NOT NULL,
+    "rent" REAL NOT NULL,
+    "team_id" VARCHAR(50)  UNIQUE REFERENCES "team" ("name") ON DELETE SET NULL
+);
 CREATE TABLE "sometable_self" (
     "backward_sts" INT NOT NULL REFERENCES "sometable" ("sometable_id") ON DELETE CASCADE,
     "sts_forward" INT NOT NULL REFERENCES "sometable" ("sometable_id") ON DELETE CASCADE
@@ -189,7 +214,7 @@ CREATE TABLE "teamevents" (
 
     async def test_schema_safe(self):
         self.maxDiff = None
-        await self.init_for("tests.models_schema_create")
+        await self.init_for("tests.schema.models_schema_create")
         sql = get_schema_sql(Tortoise.get_connection("default"), safe=True)
         self.assertEqual(
             sql.strip(),
@@ -219,6 +244,12 @@ CREATE TABLE IF NOT EXISTS "team" (
 ) /* The TEAMS! */;
 CREATE INDEX IF NOT EXISTS "idx_team_manager_676134" ON "team" ("manager_id", "key");
 CREATE INDEX IF NOT EXISTS "idx_team_manager_ef8f69" ON "team" ("manager_id", "name");
+CREATE TABLE IF NOT EXISTS "teamaddress" (
+    "city" VARCHAR(50) NOT NULL  /* City */,
+    "country" VARCHAR(50) NOT NULL  /* Country */,
+    "street" VARCHAR(128) NOT NULL  /* Street Address */,
+    "team_id" VARCHAR(50) NOT NULL UNIQUE PRIMARY KEY REFERENCES "team" ("name") ON DELETE CASCADE
+);
 CREATE TABLE IF NOT EXISTS "tournament" (
     "tid" INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
     "name" VARCHAR(100) NOT NULL  /* Tournament name */,
@@ -233,9 +264,16 @@ CREATE TABLE IF NOT EXISTS "event" (
     "token" VARCHAR(100) NOT NULL UNIQUE /* Unique token */,
     "key" VARCHAR(100) NOT NULL,
     "tournament_id" SMALLINT NOT NULL REFERENCES "tournament" ("tid") ON DELETE CASCADE /* FK to tournament */,
-    UNIQUE ("name", "prize"),
-    UNIQUE ("tournament_id", "key")
+    CONSTRAINT "uid_event_name_c6f89f" UNIQUE ("name", "prize"),
+    CONSTRAINT "uid_event_tournam_a5b730" UNIQUE ("tournament_id", "key")
 ) /* This table contains a list of all the events */;
+CREATE TABLE IF NOT EXISTS "venueinformation" (
+    "id" INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
+    "name" VARCHAR(128) NOT NULL,
+    "capacity" INT NOT NULL,
+    "rent" REAL NOT NULL,
+    "team_id" VARCHAR(50)  UNIQUE REFERENCES "team" ("name") ON DELETE SET NULL
+);
 CREATE TABLE IF NOT EXISTS "sometable_self" (
     "backward_sts" INT NOT NULL REFERENCES "sometable" ("sometable_id") ON DELETE CASCADE,
     "sts_forward" INT NOT NULL REFERENCES "sometable" ("sometable_id") ON DELETE CASCADE
@@ -317,7 +355,7 @@ class TestGenerateSchemaMySQL(TestGenerateSchema):
 
     async def test_schema(self):
         self.maxDiff = None
-        await self.init_for("tests.models_schema_create")
+        await self.init_for("tests.schema.models_schema_create")
         sql = get_schema_sql(Tortoise.get_connection("default"), safe=False)
         self.assertEqual(
             sql.strip(),
@@ -349,6 +387,13 @@ CREATE TABLE `team` (
     KEY `idx_team_manager_676134` (`manager_id`, `key`),
     KEY `idx_team_manager_ef8f69` (`manager_id`, `name`)
 ) CHARACTER SET utf8mb4 COMMENT='The TEAMS!';
+CREATE TABLE `teamaddress` (
+    `city` VARCHAR(50) NOT NULL  COMMENT 'City',
+    `country` VARCHAR(50) NOT NULL  COMMENT 'Country',
+    `street` VARCHAR(128) NOT NULL  COMMENT 'Street Address',
+    `team_id` VARCHAR(50) NOT NULL UNIQUE PRIMARY KEY,
+    CONSTRAINT `fk_teamaddr_team_1c78d737` FOREIGN KEY (`team_id`) REFERENCES `team` (`name`) ON DELETE CASCADE
+) CHARACTER SET utf8mb4;
 CREATE TABLE `tournament` (
     `tid` SMALLINT NOT NULL PRIMARY KEY AUTO_INCREMENT,
     `name` VARCHAR(100) NOT NULL  COMMENT 'Tournament name',
@@ -363,10 +408,18 @@ CREATE TABLE `event` (
     `token` VARCHAR(100) NOT NULL UNIQUE COMMENT 'Unique token',
     `key` VARCHAR(100) NOT NULL,
     `tournament_id` SMALLINT NOT NULL COMMENT 'FK to tournament',
-    UNIQUE (`name`, `prize`),
-    UNIQUE (`tournament_id`, `key`),
+    UNIQUE KEY `uid_event_name_c6f89f` (`name`, `prize`),
+    UNIQUE KEY `uid_event_tournam_a5b730` (`tournament_id`, `key`),
     CONSTRAINT `fk_event_tourname_51c2b82d` FOREIGN KEY (`tournament_id`) REFERENCES `tournament` (`tid`) ON DELETE CASCADE
 ) CHARACTER SET utf8mb4 COMMENT='This table contains a list of all the events';
+CREATE TABLE `venueinformation` (
+    `id` INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    `name` VARCHAR(128) NOT NULL,
+    `capacity` INT NOT NULL,
+    `rent` DOUBLE NOT NULL,
+    `team_id` VARCHAR(50)  UNIQUE,
+    CONSTRAINT `fk_venueinf_team_198af929` FOREIGN KEY (`team_id`) REFERENCES `team` (`name`) ON DELETE SET NULL
+) CHARACTER SET utf8mb4;
 CREATE TABLE `sometable_self` (
     `backward_sts` INT NOT NULL,
     `sts_forward` INT NOT NULL,
@@ -390,7 +443,7 @@ CREATE TABLE `teamevents` (
 
     async def test_schema_safe(self):
         self.maxDiff = None
-        await self.init_for("tests.models_schema_create")
+        await self.init_for("tests.schema.models_schema_create")
         sql = get_schema_sql(Tortoise.get_connection("default"), safe=True)
 
         self.assertEqual(
@@ -423,6 +476,13 @@ CREATE TABLE IF NOT EXISTS `team` (
     KEY `idx_team_manager_676134` (`manager_id`, `key`),
     KEY `idx_team_manager_ef8f69` (`manager_id`, `name`)
 ) CHARACTER SET utf8mb4 COMMENT='The TEAMS!';
+CREATE TABLE IF NOT EXISTS `teamaddress` (
+    `city` VARCHAR(50) NOT NULL  COMMENT 'City',
+    `country` VARCHAR(50) NOT NULL  COMMENT 'Country',
+    `street` VARCHAR(128) NOT NULL  COMMENT 'Street Address',
+    `team_id` VARCHAR(50) NOT NULL UNIQUE PRIMARY KEY,
+    CONSTRAINT `fk_teamaddr_team_1c78d737` FOREIGN KEY (`team_id`) REFERENCES `team` (`name`) ON DELETE CASCADE
+) CHARACTER SET utf8mb4;
 CREATE TABLE IF NOT EXISTS `tournament` (
     `tid` SMALLINT NOT NULL PRIMARY KEY AUTO_INCREMENT,
     `name` VARCHAR(100) NOT NULL  COMMENT 'Tournament name',
@@ -437,10 +497,18 @@ CREATE TABLE IF NOT EXISTS `event` (
     `token` VARCHAR(100) NOT NULL UNIQUE COMMENT 'Unique token',
     `key` VARCHAR(100) NOT NULL,
     `tournament_id` SMALLINT NOT NULL COMMENT 'FK to tournament',
-    UNIQUE (`name`, `prize`),
-    UNIQUE (`tournament_id`, `key`),
+    UNIQUE KEY `uid_event_name_c6f89f` (`name`, `prize`),
+    UNIQUE KEY `uid_event_tournam_a5b730` (`tournament_id`, `key`),
     CONSTRAINT `fk_event_tourname_51c2b82d` FOREIGN KEY (`tournament_id`) REFERENCES `tournament` (`tid`) ON DELETE CASCADE
 ) CHARACTER SET utf8mb4 COMMENT='This table contains a list of all the events';
+CREATE TABLE IF NOT EXISTS `venueinformation` (
+    `id` INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    `name` VARCHAR(128) NOT NULL,
+    `capacity` INT NOT NULL,
+    `rent` DOUBLE NOT NULL,
+    `team_id` VARCHAR(50)  UNIQUE,
+    CONSTRAINT `fk_venueinf_team_198af929` FOREIGN KEY (`team_id`) REFERENCES `team` (`name`) ON DELETE SET NULL
+) CHARACTER SET utf8mb4;
 CREATE TABLE IF NOT EXISTS `sometable_self` (
     `backward_sts` INT NOT NULL,
     `sts_forward` INT NOT NULL,
@@ -509,7 +577,7 @@ class TestGenerateSchemaPostgresSQL(TestGenerateSchema):
 
     async def test_schema(self):
         self.maxDiff = None
-        await self.init_for("tests.models_schema_create")
+        await self.init_for("tests.schema.models_schema_create")
         sql = get_schema_sql(Tortoise.get_connection("default"), safe=False)
         self.assertEqual(
             sql.strip(),
@@ -541,6 +609,15 @@ CREATE INDEX "idx_team_manager_676134" ON "team" ("manager_id", "key");
 CREATE INDEX "idx_team_manager_ef8f69" ON "team" ("manager_id", "name");
 COMMENT ON COLUMN "team"."name" IS 'The TEAM name (and PK)';
 COMMENT ON TABLE "team" IS 'The TEAMS!';
+CREATE TABLE "teamaddress" (
+    "city" VARCHAR(50) NOT NULL,
+    "country" VARCHAR(50) NOT NULL,
+    "street" VARCHAR(128) NOT NULL,
+    "team_id" VARCHAR(50) NOT NULL UNIQUE PRIMARY KEY REFERENCES "team" ("name") ON DELETE CASCADE
+);
+COMMENT ON COLUMN "teamaddress"."city" IS 'City';
+COMMENT ON COLUMN "teamaddress"."country" IS 'Country';
+COMMENT ON COLUMN "teamaddress"."street" IS 'Street Address';
 CREATE TABLE "tournament" (
     "tid" SMALLSERIAL NOT NULL PRIMARY KEY,
     "name" VARCHAR(100) NOT NULL,
@@ -558,13 +635,20 @@ CREATE TABLE "event" (
     "token" VARCHAR(100) NOT NULL UNIQUE,
     "key" VARCHAR(100) NOT NULL,
     "tournament_id" SMALLINT NOT NULL REFERENCES "tournament" ("tid") ON DELETE CASCADE,
-    UNIQUE ("name", "prize"),
-    UNIQUE ("tournament_id", "key")
+    CONSTRAINT "uid_event_name_c6f89f" UNIQUE ("name", "prize"),
+    CONSTRAINT "uid_event_tournam_a5b730" UNIQUE ("tournament_id", "key")
 );
 COMMENT ON COLUMN "event"."id" IS 'Event ID';
 COMMENT ON COLUMN "event"."token" IS 'Unique token';
 COMMENT ON COLUMN "event"."tournament_id" IS 'FK to tournament';
 COMMENT ON TABLE "event" IS 'This table contains a list of all the events';
+CREATE TABLE "venueinformation" (
+    "id" SERIAL NOT NULL PRIMARY KEY,
+    "name" VARCHAR(128) NOT NULL,
+    "capacity" INT NOT NULL,
+    "rent" DOUBLE PRECISION NOT NULL,
+    "team_id" VARCHAR(50)  UNIQUE REFERENCES "team" ("name") ON DELETE SET NULL
+);
 CREATE TABLE "sometable_self" (
     "backward_sts" INT NOT NULL REFERENCES "sometable" ("sometable_id") ON DELETE CASCADE,
     "sts_forward" INT NOT NULL REFERENCES "sometable" ("sometable_id") ON DELETE CASCADE
@@ -583,7 +667,7 @@ COMMENT ON TABLE "teamevents" IS 'How participants relate';
 
     async def test_schema_safe(self):
         self.maxDiff = None
-        await self.init_for("tests.models_schema_create")
+        await self.init_for("tests.schema.models_schema_create")
         sql = get_schema_sql(Tortoise.get_connection("default"), safe=True)
         self.assertEqual(
             sql.strip(),
@@ -615,6 +699,15 @@ CREATE INDEX IF NOT EXISTS "idx_team_manager_676134" ON "team" ("manager_id", "k
 CREATE INDEX IF NOT EXISTS "idx_team_manager_ef8f69" ON "team" ("manager_id", "name");
 COMMENT ON COLUMN "team"."name" IS 'The TEAM name (and PK)';
 COMMENT ON TABLE "team" IS 'The TEAMS!';
+CREATE TABLE IF NOT EXISTS "teamaddress" (
+    "city" VARCHAR(50) NOT NULL,
+    "country" VARCHAR(50) NOT NULL,
+    "street" VARCHAR(128) NOT NULL,
+    "team_id" VARCHAR(50) NOT NULL UNIQUE PRIMARY KEY REFERENCES "team" ("name") ON DELETE CASCADE
+);
+COMMENT ON COLUMN "teamaddress"."city" IS 'City';
+COMMENT ON COLUMN "teamaddress"."country" IS 'Country';
+COMMENT ON COLUMN "teamaddress"."street" IS 'Street Address';
 CREATE TABLE IF NOT EXISTS "tournament" (
     "tid" SMALLSERIAL NOT NULL PRIMARY KEY,
     "name" VARCHAR(100) NOT NULL,
@@ -632,13 +725,20 @@ CREATE TABLE IF NOT EXISTS "event" (
     "token" VARCHAR(100) NOT NULL UNIQUE,
     "key" VARCHAR(100) NOT NULL,
     "tournament_id" SMALLINT NOT NULL REFERENCES "tournament" ("tid") ON DELETE CASCADE,
-    UNIQUE ("name", "prize"),
-    UNIQUE ("tournament_id", "key")
+    CONSTRAINT "uid_event_name_c6f89f" UNIQUE ("name", "prize"),
+    CONSTRAINT "uid_event_tournam_a5b730" UNIQUE ("tournament_id", "key")
 );
 COMMENT ON COLUMN "event"."id" IS 'Event ID';
 COMMENT ON COLUMN "event"."token" IS 'Unique token';
 COMMENT ON COLUMN "event"."tournament_id" IS 'FK to tournament';
 COMMENT ON TABLE "event" IS 'This table contains a list of all the events';
+CREATE TABLE IF NOT EXISTS "venueinformation" (
+    "id" SERIAL NOT NULL PRIMARY KEY,
+    "name" VARCHAR(128) NOT NULL,
+    "capacity" INT NOT NULL,
+    "rent" DOUBLE PRECISION NOT NULL,
+    "team_id" VARCHAR(50)  UNIQUE REFERENCES "team" ("name") ON DELETE SET NULL
+);
 CREATE TABLE IF NOT EXISTS "sometable_self" (
     "backward_sts" INT NOT NULL REFERENCES "sometable" ("sometable_id") ON DELETE CASCADE,
     "sts_forward" INT NOT NULL REFERENCES "sometable" ("sometable_id") ON DELETE CASCADE