+++ tests/test_update.py
@@ -1,4 +1,4 @@
-from tests.testmodels import Contact, ContactTypeEnum, Event, Tournament
+from tests.testmodels import Event, Tournament
 from tortoise.contrib import test
 
 
@@ -18,15 +18,3 @@ class TestUpdate(test.TestCase):
         await Event.all().update(tournament=tournament_second)
         event = await Event.first()
         self.assertEqual(event.tournament_id, tournament_second.id)
-
-    async def test_update_with_int_enum_value(self):
-        contact = await Contact.create()
-        contact.type = ContactTypeEnum.home
-        await contact.save()
-
-    async def test_exception_on_invalid_data_type_in_int_field(self):
-        contact = await Contact.create()
-
-        contact.type = "not_int"
-        with self.assertRaises((TypeError, ValueError)):
-            await contact.save()