+++ tortoise/backends/sqlite/executor.py
@@ -26,14 +26,16 @@ def to_db_decimal(self, value, instance) -> Optional[str]:
 
 
 def to_db_datetime(self, value: Optional[datetime.datetime], instance) -> Optional[str]:
-    if self.auto_now:
-        value = datetime.datetime.utcnow()
-        setattr(instance, self.model_field_name, value)
-        return value.isoformat(" ")
-    if self.auto_now_add and getattr(instance, self.model_field_name) is None:
-        value = datetime.datetime.utcnow()
-        setattr(instance, self.model_field_name, value)
-        return value.isoformat(" ")
+    if hasattr(instance, "_saved_in_db"):
+        # Only do this if it is a Model instance, not class. Test for guaranteed instance var
+        if self.auto_now:
+            value = datetime.datetime.utcnow()
+            setattr(instance, self.model_field_name, value)
+            return value.isoformat(" ")
+        if self.auto_now_add and getattr(instance, self.model_field_name, None) is None:
+            value = datetime.datetime.utcnow()
+            setattr(instance, self.model_field_name, value)
+            return value.isoformat(" ")
     if isinstance(value, datetime.datetime):
         return value.isoformat(" ")
     return None