+++ tortoise/queryset.py
@@ -562,7 +562,7 @@ class UpdateQuery(AwaitableQuery):
                 raise FieldError(f"Unknown keyword argument {key} for model {self.model}")
             if field_object.pk:
                 raise IntegrityError(f"Field {key} is PK and can not be updated")
-            if isinstance(field_object, fields.ForeignKeyField):
+            if isinstance(field_object, (fields.ForeignKeyField, fields.OneToOneField)):
                 fk_field: str = field_object.source_field  # type: ignore
                 db_field = self.model._meta.fields_map[fk_field].source_field
                 value = executor.column_map[fk_field](value.pk, None)