+++ CHANGELOG.rst
@@ -2,6 +2,38 @@
 
 Changelog
 =========
+0.15.4
+------
+- Don't generate a schema if there is no models (#254)
+- Emit a ``RuntimeWarning`` when a module has no models to import (#254)
+- Allow passing in a custom SSL context (#255)
+
+0.15.3
+------
+* Added ``OneToOneField`` implementation:
+
+  ``OneToOneField`` describes a one to one relation between two models.
+  It can be set from the primary side only, but resolved from both sides in the same way.
+
+  ``describe_model(...)`` now also reports OneToOne relations in both directions.
+
+  Usage example:
+
+  .. code-block:: python3
+
+     event: fields.OneToOneRelation[Event] = fields.OneToOneField(
+         "models.Event", on_delete=fields.CASCADE, related_name="address"
+     )
+
+- Prefetching is done concurrently now, sending all prefetch requests at the same time instead of in sequence.
+- Enabe foreign key enforcement on SQLite for builds where it was optional.
+
+0.15.2
+------
+- The ``auto_now_add`` argument of ``DatetimeField`` is handled correctly in the SQLite backend.
+- ``unique_together`` now creates named constrains, to prevent the DB from auto-assigning a potentially non-unique constraint name.
+- Filtering by an ``auto_now`` field doesn't replace the filter value with ``now()`` anymore.
+
 0.15.1
 ------
 - Handle OR'ing a blank ``Q()`` correctly (#240)